Build with embedded pjproject 2.8 (dfsg-repacked)
Upstream only supports a private build of pjproject for Asterisk 16 anymore.
Since we cannot download it during build and the original tarball is not
DFSG-free rebuild pjproject tarball using mk-origtargz and a copy of
debian/copyright from src:pjproject. Ship it in debian/

Documented in debian/README.source

Drop build-dep on libpjproject-dev